+++ src/Symfony/Component/Mailer/Bridge/Sendgrid/Tests/Transport/SendgridApiTransportTest.php
@@ -244,4 +244,21 @@ class SendgridApiTransportTest extends TestCase
         $this->assertSame('blue', $payload['personalizations'][0]['custom_args']['Color']);
         $this->assertSame('12345', $payload['personalizations'][0]['custom_args']['Client-ID']);
     }
+
+    public function testEnvelopeOptions()
+    {
+        $expectedMailerSettings = ['bypass_unsubscribe_management' => ['enable' => true]];
+
+        $email = new Email();
+        $envelope = new Envelope(new Address('envelopefrom@example.com'), [new Address('envelopeto@example.com')]);
+        $envelope->setOption('mail_settings', $expectedMailerSettings);
+
+        $transport = new SendgridApiTransport('ACCESS_KEY');
+        $method = new \ReflectionMethod(SendgridApiTransport::class, 'getPayload');
+        $method->setAccessible(true);
+        $payload = $method->invoke($transport, $email, $envelope);
+
+        $this->assertArrayHasKey('mail_settings', $payload);
+        $this->assertSame($payload['mail_settings'], $expectedMailerSettings);
+    }
 }